Rserve authentication working
The Username and Password fields in GNU R Options Panel are now
connected to the backend and thus the authentication ist working.
Password is stored in the Keyring provided by the operating system. This
is currently tested under windows only, linux and Mac still need
testing. The Option panel still needs some fine tuning as the "apply"
button is not always enebled if something is changed.