Completed the implementation of a List based TableModel and some need helper classes.
Additional classed had to be modified in order to work with the new List
based Table Model. These classes are copies of their Vector based
original with the word "List" appended to their names. For interfaces
the word "List" got appended before the "I" at the end of the class
name. A first place where this new List based Table Model is used is the
ResultDeAnalysis class. When beeing used at other places in the
application it might be possible that additional helper classes need
modification.