introduce ReadXplorerException
ReadXplorerException is an abstract RuntimeException which is a base class for problem centric Exception which should be handled over from models/controllers to UIs.
Thus, every public method of a model/controller should throw a specific version of ReadXplorerException. Hereby, UIs CAN catch them where appropriate but don't have to!